Rewrite `range` to `exists` query when bounds are open (22640)
Range queries that have both bounds open are costly to execute. The rewrite phase of queries
provides the opportunity to rewrite queries as faster running simpler queries. This PR switches to
an `exists` queries if both bounds are open.

The output should be the same for both queries below, but under the hood an exists query is ultimately used

```java
// typical query
GET _search
{
    "query" : {
        "range" : {
            "post_date" : { "from" : "2009-11-15T13:00:00", "to" : "2009-11-15T14:00:00" }
        }
    }
}'
```

```java
// open bounds query
GET _search
{
    "query" : {
        "range" : {
            "post_date" : { }
        }
    }
}'
```

Closes issue 22640