Issue:21802 - Squashed Commits(3 Commits)
Functionality changes: Fixing a bug.
The issue announced that 'require_field_match' does not accept String
value "true" and "false", although they have the same meaning as boolean
value in a query.
This commit fixed this problem by modifying exception mechanism. Logics
have been added to ObjectParser.java in order to let String value "true"
and "false" pass but not throw an exception.

Issue:21802 - Add comments

A few comments includes introduction of modification on the code,
mechanism and logics which are relevant to highlight searching function.
This is to explain how the issue is fixed and give a brief instruct on
how the function works.

Issue:21802 - Refactoring

This commit :
1.Consider that method assertSupports might be called by other object in
the project, so delete the formal parameter which has just been added to
method assertSupport. Instead, define a global variable in the super
class AbstractObjectParser. When another query calls parse method, this
parser of that query will be assigned to this variable. The object of
XContentParser is very useful for acquiring current inputted commands.
2. Put the static blocks to the top in HighlightBuilder and its abstract
superclass, together with some doc introduction. This is to inform the
future contributors that these blocks loaded in highest prority to read
inputted value into memory and check them before commencing highlight
search logic.