Issue:21802 - Refactoring
This commit :
1.Consider that method assertSupports might be called by other object in
the project, so delete the formal parameter which has just been added to
method assertSupport. Instead, define a global variable in the super
class AbstractObjectParser. When another query calls parse method, this
parser of that query will be assigned to this variable. The object of
XContentParser is very useful for acquiring current inputted commands.
2. Put the static blocks to the top in HighlightBuilder and its abstract
superclass, together with some doc introduction. This is to inform the
future contributors that these blocks loaded in highest prority to read
inputted value into memory and check them before commencing highlight
search logic.